requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf
langchain-ollama
